+++ anymotion_cli/commands/analyze.py
@@ -3,9 +3,9 @@ from typing import Optional
 
 import click
 from anymotion_sdk import RequestsError
-from click_help_colors import HelpColorsGroup
 from yaspin import yaspin
 
+from ..click_custom import CustomCommand
 from ..exceptions import ClickException
 from ..options import common_options
 from ..output import echo, echo_success
@@ -14,16 +14,23 @@ from ..utils import color_id, get_client, parse_rule
 from .analysis import show
 
 
-@click.group(cls=HelpColorsGroup, help_options_color="cyan")
+@click.group()
 def cli() -> None:  # noqa: D103
     pass
 
 
-@cli.command(short_help="Analyze the extracted keypoint data.")
+@cli.command(
+    cls=CustomCommand,
+    help_options_color="cyan",
+    short_help="Analyze the extracted keypoint data.",
+)
 @click.argument("keypoint_id", type=int)
 @click.option("--rule", "rule_str", help="Analysis rules in JSON format.")
 @click.option(
-    "--rule-file", type=click.File(), help="Analysis rules file in JSON format."
+    "--rule-file",
+    type=click.File(),
+    metavar="PATH",
+    help="Path of analysis rules file in JSON format.",
 )
 @click.option("--show-result", is_flag=True)
 @common_options