+++ anymotion_cli/exceptions.py
@@ -21,3 +21,15 @@ class SettingsException(Exception):
 
 class SettingsValueError(SettingsException):
     """Raised when settings value is invalid."""
+
+
+class HelpColorsException(Exception):
+    """Raised when unknown color is given."""
+
+
+class InternalCommandException(Exception):
+    """Base class for exceptions in the internal exception."""
+
+
+class ExitReplException(InternalCommandException):
+    """Exit repl."""